refactor: 핸들러 추상클래스 이동
